Harmonize naming of tests
